Fix bug where PHP error reporting was always zero.
When the error_reporting directive is set through nginx, the directive
is not parsed through PHP, so setting it to "E_ALL" was actually
translating to "0" instead of "32767". Passing integers fixes this
problem. See <http://php.net/manual/en/errorfunc.constants.php> for more
info on values to pass instead of the constant names.